+++ example/add_box.py
@@ -1,3 +1,5 @@
+"""在 PlateCarree 和 AzimuthalEquidistant 投影的地图上画方块格子"""
+
 import matplotlib.pyplot as plt
 import numpy as np