+++ example/chongqing.py
@@ -1,3 +1,5 @@
+"""高德数据在市级行政区划上将重庆分为了重庆城区和重庆郊县，展示其划分"""
+
 import matplotlib.pyplot as plt
 from matplotlib.patches import Patch