+++ src/frykit/shp/data.py
@@ -11,7 +11,7 @@ import pandas as pd
 import shapely
 
 from frykit import get_data_dirpath
-from frykit._config import DataSource, config
+from frykit.configuration import DataSource, config
 from frykit.shp.binary import BinaryReader
 from frykit.shp.typing import LineStringType, PolygonType
 from frykit.utils import deprecator, format_type_error, to_list